requirements.txt
asgiref==3.8.1
Django==4.2.20
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
numpy==2.0.2
opencv-python==4.11.0.86
pillow==11.2.1
PyJWT==2.9.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
